Phpfloat used now double type

+++ src/Type/PhpFloat.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Cassandra\Type;
 
 /**
- * @deprecated Use \Cassandra\Type\Float32 instead
+ * @deprecated Use \Cassandra\Type\Double instead
  */
-final class PhpFloat extends Float32 {
+final class PhpFloat extends Double {
 }
